==> 3_split_by_polymer.R <==
t <- read_csv("./new/tidyPE462_abundances_calc_v5.csv") ######file paths might need to be corrected
contr <- t %>% filter(polymer %in% c("Glass_fiber","mock_DNA","Neg_PCR"))
unique(contr$polymer)
write_csv(contr, "./new/PE462_controls_v5.csv")
pet <- rbind((t %>% filter(polymer == "PET" )), (t %>% filter(material =="wood")))
unique(pet$polymer)
write_csv(pet, "./new/PE462_PET_v5.csv")
pe <- rbind((t %>% filter(polymer == "PE" )), (t %>% filter(material =="wood")))
unique(pe$polymer)
write_csv(pe, "./new/PE462_PE_v5.csv")
ps <- rbind((t %>% filter(polymer == "PS" )), (t %>% filter(material =="wood")))
unique(ps$polymer)
write_csv(ps, "./new/PE462_PS_v5.csv")
Nylon <- rbind((t %>% filter(polymer == "Nylon" )), (t %>% filter(material =="wood")))
unique(Nylon$polymer)
write_csv(Nylon, "./new/PE462_Nylon_5.csv")
